Routeur pour accueillir le calculateur inversé
